requirements.txt
streamlit>=1.28.0
langchain-ollama>=0.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pandas>=1.5.0
plotly>=5.0.0
pypdf>=3.0.0
pyyaml>=6.0
numpy>=1.21.0


faiss-gpu>=1.7.0


pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950